miRkwood: Update files for unit tests.
Now we use the basic_candidates.yml file to create the results
files so I needed to create a basic_candidates.yml file.

The alternative candidates in the candidate YML file have now
the MFEI concatenated with the name in the name field.
I could delete the other features and keep only the sequence
and the structure stemloop since they are the only needed,
I'm just lazy here.