miRkwood: Delete useless style properties.
We don't need to specify a font family, it is already defined
in bioinfo.css (common style for all softwares).

We don't need to speciify a font size, it is overrided by
bioinfo.css except for the title (the font size is specified
for id main), so on pages using help.css previously we had
smaller titles. Now all titles are the same.