miRkwood: Web smallRNA-seq pipeline: add a column 'OCC.' for the nb of candidates with the same read as miRNA.
In the basic_candidates.yml file, for each candidate the array
contains the ID of all OTHER candidates with the same miRNA
(no self reference), in the html table we show the nb of
candidates with the same miRNA, ie we add 1 to the array
length. 
First we need to check if there is a miRNA or not (if no
miRNA, the column stays empty). If there is a miRNA and
the array is empty, it means that this candidate is the only
one with this miRNA so the column contains 1.