miRkwood: detailed page for miRNAs: use the same page structure for the 3 cases (valid ID and some results, valid ID and no results, and no valid ID).
The only changing part is in the <div id="main"> </div>
so we store this part in a new variable and we make the
page only once, at the end.

Theoretically, if there is no results or if the ID is not
valid, this page is not called, but we never know... If
the user want to modify the URL to access this page, the
page structure will be good!