miRkwood: Change the way we merge the alignments.
If there are several alignments for a candidate, we merge
all alignments [c;d] in alignment [a;b] if a <= c and b >= d+2.

Previously we kept as the key for the merged positions the
position that has the most hits.

Now we don't keep the details about alignments, we only need
the positions, and in case of multiple alignments, we keep 
the larger positions of the alignment. 
Example: we merge the alignments [5;24] and [6;26] (since
5 <= 6 and 26 >= 24+2), the key will become [5;26].
Previously it was [5;24]. But in the multiple alignment file
created by Alan the positions were already "merged" into
[5;26] (it is necessary to be able to show all alignments)
but the miRbase tags on the reads clouds are based on the
positions in the YML file and the colored hairpin (HTML file)
is based on the alignment file so we needed to have the same
positions, elsewhere the information about the alignments
would not be consistent.