miRkwood: Web pipelines : add PDF in the list of available formats for exports.
The ORG file is needed to create the PDF file, so we create
it first.
Now the ORG file path is an attribute of class PDFExporter,
instead of being called as an attribute for function 
create_PDF_from_ORG().

Both ORG and PDF files are stored in /tmp. They will be
created each time we export results in PDF format.