miRkwood: First attempt of a Dockerfile.
Inline::CPP is still unpredictable when it comes to
installation. I had problems while installing some
dependencies (ExtUtils::CppGuess, Inline::C).
It seems to be OK if we use --force option, but it's
weird since last week on an unversioned (shame on me)
version of the Dockerfile, it worked very well without
'--force'.

For some reason it seems that making symbolic links in
the docker is a bad idea. RNAstemloop was not a "plain
file" anymore so it couldn't run. To solve this problem,
since we choose the architecture, in docker we edit
the programs.cfg file so that miRkwood knows it must use
RNAstemloop-x86_64.

To run mirkwood in Docker the command line is something like
that :
(sudo) docker run 
-v /input_path_on_your_machine/:/input_path_in_docker/
-v /output_path_on_your_machine/:/output_path_in_docker/
myImageName 
perl -I/home/mirkwood/cgi-bin/lib/ /home/mirkwood/cgi-bin/bin/mirkwood.pl
--input /input_path_in_docker/file.fasta (or bed) 
--output /output_path_in_docker/output_dir (other options)

You can map as many folders as you want with the -v option, 
you can also put input and output files and directories
in an unique folder for the mapping.

myImageName is the name of the mirkwood docker image.