miRkwood: HTML results page (CLI smallRNAseq pipeline, new candidates): change the header.
Now we have a part for the candidate features (chromosome, position,
strand, nb of reads, reads distribution score, mfei and optionally 
shuffles) and a part for the mirna features(sequence, length, weight
and optionally alignment).
miRNA features have an additional header line (with just "miRNA")
while candidate features are in cells that spread on 2 lines.