miRkwood: add a safety about filtering bed.
The gff file used with bedtools intersect may be readable
and still present problems.
Now we check the result of the "system" command.
If it's non zero, the output file is copied from the input file
(before that, we got empty output files when there was a problem
with the GFF file)